+++ frontend/src/App.css
@@ -1,0 +1,233 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans:wght@400;500;600&display=swap');
+
+/* remove the scrollbar from inputs */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+:root {
+  --primary: #6119d6;
+  --secondary: #f7f322;
+  --bg-color: #ffffffff;
+  --font-color: white;
+}
+
+
+.App {
+  text-align: center;
+}
+
+body{
+  /* background-color: var(--bg-color); */
+  background-color: var(--primary );
+  
+}
+
+.container{
+  padding: 50px 30px 70px 30px;
+  border-radius: 15px;
+  margin-top: 0px;
+  font-size: 19px;
+}
+
+.allInputs{
+  color: var(--font-color);
+  margin-left: 180px !important;
+  width: calc(90vw - 300px) !important;
+}
+
+.btn{
+  font-family: inherit;
+  padding: 20px;
+  border-radius: 10px;
+  border: none;
+  margin-top: 20px;
+  font-weight: 500;
+  font-size: 13px;
+  cursor: pointer;
+}
+
+.accountBalance{
+  color: white;
+  font-size: 20px;
+}
+
+.userOutput{
+  margin-top: 50px;
+  font-size: 20px;
+  color:yellow;
+}
+
+.owner{
+  width: calc(90vw -300px);
+  font-family:'Noto Sans', sans-serif;
+  color: var(--font-color);
+  margin-left: 320px;
+  margin-top: 20px;
+}
+
+.setAllowance{
+  padding-top: 10px;
+  padding-bottom: 15px;
+}
+
+.owner-btn{
+  margin-right: 10px;
+  background-color: var(--secondary);
+  color: black;
+  width: 200px;
+  font-family: inherit;
+}
+
+.txtbox{
+  font-family: inherit;
+  font-size: inherit;
+  background-color: var(--primary);
+  border: 2px solid white;
+  margin-right: 60px;
+  margin-left: 15px;
+  height: 25px;
+  width: 300px;
+  border-radius: 5px;
+}
+
+.txtbox:hover{
+  border: 2px solid yellow;
+  cursor: pointer;
+}
+
+.txtbox:focus{
+  border: 2px solid yellow;
+  cursor: pointer; 
+  background-color: rgb(194, 189, 189);
+}
+
+.btn:hover{
+  cursor: pointer;
+  background-color: #c09f0c;
+  box-shadow: 2px 2px 5px rgb(53, 50, 50);
+}
+
+/* --- NAV STYLES --- */
+.sidenav {
+  padding-top: 60px !important;
+  height: 100vh;
+  width: 300px;
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  background-color: #111;
+  overflow-x: hidden;
+  padding-top: 20px;
+}
+
+.sidenav a {
+  padding: 10px 10px 10px 32px;
+  text-decoration: none;
+  font-size: 25px;
+  color: #818181;
+  display: block;
+}
+
+.sidenav a:hover {
+  color: #f1f1f1;
+  background-color: #353535;
+}
+.sidenav a.active{
+  background-color: var(--primary);
+  color: white;
+}
+
+
+
+
+
+/* -- stake input styles -- */
+.checkbox{
+  padding-top: 50px;
+}
+.checkbox legend{
+  padding-bottom: 20px;
+}
+.radioChoices label{
+  padding-right: 30px;
+}
+input[type="radio"] {
+  display:none;
+}
+
+input[type="radio"] + label {
+  color:#f2f2f2;
+  font-family:Arial, sans-serif;
+}
+
+input[type="radio"] + label span {
+  display:inline-block;
+  width:19px;
+  height:19px;
+  margin:-2px 10px 0 0;
+  vertical-align:middle;
+  background:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/210284/check_radio_sheet.png) -38px top no-repeat;
+  cursor:pointer;
+}
+
+input[type="radio"]:checked + label span {
+  background:url(https://s3-us-west-2.amazonaws.com/s.cdpn.io/210284/check_radio_sheet.png) -57px top no-repeat;
+}
+
+
+.stakeContainer{
+  display: flex;
+  border: white 2px solid;
+  padding: 30px 10px 10px 30px;
+  width: 550px;
+  border-radius: 10px;
+}
+
+.stakeContainer h3{
+  padding-top: 0;
+  margin-top: 0;
+}
+
+input[type-range]{
+  -webkit-appearance: none;
+  appearance: none;
+}
+
+input[type=range]::-webkit-slider-runnable-track {
+  width: 100%;
+  height: 8px;
+  cursor: pointer;
+  box-shadow: 1px 1px 1px #000000;
+  background: #ffffff;
+  border-radius: 5px;
+}
+
+input[type=range]::-webkit-slider-thumb {
+  box-shadow: 1px 1px 1px #000000;
+  height: 30px;
+  width: 30px;
+  border-radius: 5px;
+  color: red;
+  background-color: red;
+  cursor: pointer;
+  margin-top: -5px;
+}
+input[type=range]:hover {
+  outline: none;
+}
+
+input[type=range]:hover::-webkit-slider-runnable-track {
+  background: #85909b;
+}
+
+.stakeReturnOuput{
+  display: flex;
+}
+
+.stakeReturnOuput h3{
+  padding-right: 40px;
+}
+
