--- package.json
@@ -1,42 +1,0 @@
-{
-  "name": "allowance-dapp",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@sweetalert2/themes": "^5.0.15",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "ethers": "^5.7.0",
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0",
-    "react-scripts": "5.0.1",
-    "sweetalert2": "^11.6.16",
-    "web-vitals": "^2.1.4",
-    "web3modal": "^1.9.9"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
